Fix dumb typo...

+++ src/main/java/org/metamechanists/quaptics/implementation/blocks/manipulators/Capacitor.java
@@ -76,7 +76,7 @@ public class Capacitor extends ConnectedBlock {
         }
 
         final String chargeString = BlockStorage.getLocationInfo(group.getLocation(), Keys.CHARGE);
-        double charge = chargeString == null ? 0 : Integer.parseInt(chargeString);
+        double charge = chargeString == null ? 0 : Double.parseDouble(chargeString);
         if (input.isLinkEnabled()) {
             final double additionalCharge = input.getLink().getPower() / QuapticTicker.QUAPTIC_TICKS_PER_SECOND;
             charge = settings.stepCharge(charge, additionalCharge);
